services/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
sqlalchemy==2.0.23
asyncpg==0.29.0
python-multipart==0.0.6
aiofiles==23.2.1
